ssh-key: initial `serde` support
Support for serializing/deserializing `Certificate` and `PublicKey`
using `serde`'s traits.

Uses the OpenSSH string serialization with human-readable formats, and
the binary serialization with binary formats.
