+++ ssh-key/Cargo.toml
@@ -27,6 +27,7 @@ bcrypt-pbkdf = { version = "0.9", optional = true, default-features = false }
 ed25519-dalek = { version = "1.0.1", optional = true, default-features = false, features = ["u64_backend"] }
 rand_core = { version = "0.6", optional = true, default-features = false }
 sec1 = { version = "=0.3.0-pre.1", optional = true, default-features = false, features = ["point"], path = "../sec1" }
+serde = { version = "1", optional = true }
 sha2 = { version = "0.10", optional = true, default-features = false }
 signature = { version = "1.3.1", optional = true, default-features = false }
 subtle = { version = "2", optional = true, default-features = false }
