+++ ssh-key/src/public.rs
@@ -7,39 +7,43 @@ mod dsa;
 #[cfg(feature = "ecdsa")]
 mod ecdsa;
 mod ed25519;
+mod key_data;
 mod openssh;
 #[cfg(feature = "alloc")]
 mod rsa;
 
+pub use self::{ed25519::Ed25519PublicKey, key_data::KeyData};
+
 #[cfg(feature = "ecdsa")]
 pub use self::ecdsa::EcdsaPublicKey;
-pub use self::ed25519::Ed25519PublicKey;
 #[cfg(feature = "alloc")]
 pub use self::{dsa::DsaPublicKey, rsa::RsaPublicKey};
 
 pub(crate) use self::openssh::Encapsulation;
 
 use crate::{
-    checked::CheckedSum,
     decode::Decode,
     encode::Encode,
     reader::{Base64Reader, Reader},
-    writer::Writer,
     Algorithm, Error, Result,
 };
 use core::str::FromStr;
 
 #[cfg(feature = "alloc")]
 use {
-    crate::writer::base64_len,
+    crate::{checked::CheckedSum, writer::base64_len},
     alloc::{
         borrow::ToOwned,
         string::{String, ToString},
+        vec::Vec,
     },
 };
 
 #[cfg(feature = "fingerprint")]
-use crate::{Fingerprint, HashAlg, Sha256Fingerprint};
+use crate::{Fingerprint, HashAlg};
+
+#[cfg(all(feature = "alloc", feature = "serde"))]
+use serde::{de, ser, Deserialize, Serialize};
 
 #[cfg(feature = "std")]
 use std::{fs, path::Path};
@@ -75,8 +79,8 @@ impl PublicKey {
     /// ```text
     /// ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAILM+rvN+ot98qgEN796jTiQfZfG1KaT0PtFDJ/XFSqti foo@bar.com
     /// ```
-    pub fn from_openssh(input: impl AsRef<[u8]>) -> Result<Self> {
-        let encapsulation = Encapsulation::decode(input.as_ref())?;
+    pub fn from_openssh(public_key: &str) -> Result<Self> {
+        let encapsulation = Encapsulation::decode(public_key.trim_end().as_bytes())?;
         let mut reader = Base64Reader::new(encapsulation.base64_data)?;
         let key_data = KeyData::decode(&mut reader)?;
 
@@ -94,6 +98,18 @@ impl PublicKey {
             #[cfg(feature = "alloc")]
             comment: encapsulation.comment.to_owned(),
         })
+    }
+
+    /// Parse a raw binary SSH public key.
+    pub fn from_bytes(mut bytes: &[u8]) -> Result<Self> {
+        let reader = &mut bytes;
+        let key_data = KeyData::decode(reader)?;
+
+        if reader.is_finished() {
+            Ok(key_data.into())
+        } else {
+            Err(Error::Length)
+        }
     }
 
     /// Encode OpenSSH-formatted public key.
@@ -120,6 +136,15 @@ impl PublicKey {
         let actual_len = self.encode_openssh(&mut buf)?.len();
         buf.truncate(actual_len);
         Ok(String::from_utf8(buf)?)
+    }
+
+    /// Serialize SSH public key as raw bytes.
+    #[cfg(feature = "alloc")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
+    pub fn to_bytes(&self) -> Result<Vec<u8>> {
+        let mut ret = Vec::new();
+        self.key_data.encode(&mut ret)?;
+        Ok(ret)
     }
 
     /// Read public key from an OpenSSH-formatted file.
@@ -219,184 +244,38 @@ impl ToString for PublicKey {
     }
 }
 
-/// Public key data.
-#[derive(Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
-#[non_exhaustive]
-pub enum KeyData {
-    /// Digital Signature Algorithm (DSA) public key data.
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
-    Dsa(DsaPublicKey),
-
-    /// Elliptic Curve Digital Signature Algorithm (ECDSA) public key data.
-    #[cfg(feature = "ecdsa")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "ecdsa")))]
-    Ecdsa(EcdsaPublicKey),
-
-    /// Ed25519 public key data.
-    Ed25519(Ed25519PublicKey),
-
-    /// RSA public key data.
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
-    Rsa(RsaPublicKey),
-}
-
-impl KeyData {
-    /// Get the [`Algorithm`] for this public key.
-    pub fn algorithm(&self) -> Algorithm {
-        match self {
-            #[cfg(feature = "alloc")]
-            Self::Dsa(_) => Algorithm::Dsa,
-            #[cfg(feature = "ecdsa")]
-            Self::Ecdsa(key) => key.algorithm(),
-            Self::Ed25519(_) => Algorithm::Ed25519,
-            #[cfg(feature = "alloc")]
-            Self::Rsa(_) => Algorithm::Rsa { hash: None },
-        }
-    }
-
-    /// Get DSA public key if this key is the correct type.
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
-    pub fn dsa(&self) -> Option<&DsaPublicKey> {
-        match self {
-            Self::Dsa(key) => Some(key),
-            _ => None,
-        }
-    }
-
-    /// Get ECDSA public key if this key is the correct type.
-    #[cfg(feature = "ecdsa")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "ecdsa")))]
-    pub fn ecdsa(&self) -> Option<&EcdsaPublicKey> {
-        match self {
-            Self::Ecdsa(key) => Some(key),
-            _ => None,
-        }
-    }
-
-    /// Get Ed25519 public key if this key is the correct type.
-    pub fn ed25519(&self) -> Option<&Ed25519PublicKey> {
-        match self {
-            Self::Ed25519(key) => Some(key),
-            #[allow(unreachable_patterns)]
-            _ => None,
-        }
-    }
-
-    /// Compute key fingerprint.
-    ///
-    /// Use [`Default::default()`] to use the default hash function (SHA-256).
-    #[cfg(feature = "fingerprint")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "fingerprint")))]
-    pub fn fingerprint(&self, hash_alg: HashAlg) -> Result<Fingerprint> {
-        match hash_alg {
-            HashAlg::Sha256 => Sha256Fingerprint::try_from(self).map(Into::into),
-            _ => Err(Error::Algorithm),
-        }
-    }
-
-    /// Get RSA public key if this key is the correct type.
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
-    pub fn rsa(&self) -> Option<&RsaPublicKey> {
-        match self {
-            Self::Rsa(key) => Some(key),
-            _ => None,
-        }
-    }
-
-    /// Is this key a DSA key?
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
-    pub fn is_dsa(&self) -> bool {
-        matches!(self, Self::Dsa(_))
-    }
-
-    /// Is this key an ECDSA key?
-    #[cfg(feature = "ecdsa")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "ecdsa")))]
-    pub fn is_ecdsa(&self) -> bool {
-        matches!(self, Self::Ecdsa(_))
-    }
-
-    /// Is this key an Ed25519 key?
-    pub fn is_ed25519(&self) -> bool {
-        matches!(self, Self::Ed25519(_))
-    }
-
-    /// Is this key an RSA key?
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
-    pub fn is_rsa(&self) -> bool {
-        matches!(self, Self::Rsa(_))
-    }
-
-    /// Decode [`KeyData`] for the specified algorithm.
-    pub(crate) fn decode_algorithm(reader: &mut impl Reader, algorithm: Algorithm) -> Result<Self> {
-        match algorithm {
-            #[cfg(feature = "alloc")]
-            Algorithm::Dsa => DsaPublicKey::decode(reader).map(Self::Dsa),
-            #[cfg(feature = "ecdsa")]
-            Algorithm::Ecdsa { curve } => match EcdsaPublicKey::decode(reader)? {
-                key if key.curve() == curve => Ok(Self::Ecdsa(key)),
-                _ => Err(Error::Algorithm),
-            },
-            Algorithm::Ed25519 => Ed25519PublicKey::decode(reader).map(Self::Ed25519),
-            #[cfg(feature = "alloc")]
-            Algorithm::Rsa { .. } => RsaPublicKey::decode(reader).map(Self::Rsa),
-            #[allow(unreachable_patterns)]
-            _ => Err(Error::Algorithm),
-        }
-    }
-
-    /// Get the encoded length of this key data without a leading algorithm
-    /// identifier.
-    pub(crate) fn encoded_key_data_len(&self) -> Result<usize> {
-        match self {
-            #[cfg(feature = "alloc")]
-            Self::Dsa(key) => key.encoded_len(),
-            #[cfg(feature = "ecdsa")]
-            Self::Ecdsa(key) => key.encoded_len(),
-            Self::Ed25519(key) => key.encoded_len(),
-            #[cfg(feature = "alloc")]
-            Self::Rsa(key) => key.encoded_len(),
-        }
-    }
-
-    /// Encode the key data without a leading algorithm identifier.
-    pub(crate) fn encode_key_data(&self, writer: &mut impl Writer) -> Result<()> {
-        match self {
-            #[cfg(feature = "alloc")]
-            Self::Dsa(key) => key.encode(writer),
-            #[cfg(feature = "ecdsa")]
-            Self::Ecdsa(key) => key.encode(writer),
-            Self::Ed25519(key) => key.encode(writer),
-            #[cfg(feature = "alloc")]
-            Self::Rsa(key) => key.encode(writer),
+#[cfg(all(feature = "alloc", feature = "serde"))]
+#[cfg_attr(docsrs, doc(cfg(all(feature = "alloc", feature = "serde"))))]
+impl<'de> Deserialize<'de> for PublicKey {
+    fn deserialize<D>(deserializer: D) -> core::result::Result<Self, D::Error>
+    where
+        D: de::Deserializer<'de>,
+    {
+        if deserializer.is_human_readable() {
+            let string = String::deserialize(deserializer)?;
+            Self::from_openssh(&string).map_err(de::Error::custom)
+        } else {
+            let bytes = Vec::<u8>::deserialize(deserializer)?;
+            Self::from_bytes(&bytes).map_err(de::Error::custom)
         }
     }
 }
 
-impl Decode for KeyData {
-    fn decode(reader: &mut impl Reader) -> Result<Self> {
-        let algorithm = Algorithm::decode(reader)?;
-        Self::decode_algorithm(reader, algorithm)
-    }
-}
-
-impl Encode for KeyData {
-    fn encoded_len(&self) -> Result<usize> {
-        [
-            self.algorithm().encoded_len()?,
-            self.encoded_key_data_len()?,
-        ]
-        .checked_sum()
-    }
-
-    fn encode(&self, writer: &mut impl Writer) -> Result<()> {
-        self.algorithm().encode(writer)?;
-        self.encode_key_data(writer)
+#[cfg(all(feature = "alloc", feature = "serde"))]
+#[cfg_attr(docsrs, doc(cfg(all(feature = "alloc", feature = "serde"))))]
+impl Serialize for PublicKey {
+    fn serialize<S>(&self, serializer: S) -> core::result::Result<S::Ok, S::Error>
+    where
+        S: ser::Serializer,
+    {
+        if serializer.is_human_readable() {
+            self.to_openssh()
+                .map_err(ser::Error::custom)?
+                .serialize(serializer)
+        } else {
+            self.to_bytes()
+                .map_err(ser::Error::custom)?
+                .serialize(serializer)
+        }
     }
 }
