+++ ssh-key/tests/certificate.rs
@@ -150,7 +150,7 @@ fn decode_rsa_4096_openssh() {
 #[test]
 fn encode_dsa_openssh() {
     let cert = Certificate::from_str(DSA_CERT_EXAMPLE).unwrap();
-    assert_eq!(DSA_CERT_EXAMPLE.trim_end(), &cert.to_string().unwrap());
+    assert_eq!(DSA_CERT_EXAMPLE.trim_end(), &cert.to_openssh().unwrap());
 }
 
 #[cfg(feature = "ecdsa")]
@@ -159,20 +159,23 @@ fn encode_ecdsa_p256_openssh() {
     let cert = Certificate::from_str(ECDSA_P256_CERT_EXAMPLE).unwrap();
     assert_eq!(
         ECDSA_P256_CERT_EXAMPLE.trim_end(),
-        &cert.to_string().unwrap()
+        &cert.to_openssh().unwrap()
     );
 }
 
 #[test]
 fn encode_ed25519_openssh() {
     let cert = Certificate::from_str(ED25519_CERT_EXAMPLE).unwrap();
-    assert_eq!(ED25519_CERT_EXAMPLE.trim_end(), &cert.to_string().unwrap());
+    assert_eq!(ED25519_CERT_EXAMPLE.trim_end(), &cert.to_openssh().unwrap());
 }
 
 #[test]
 fn encode_rsa_4096_openssh() {
     let cert = Certificate::from_str(RSA_4096_CERT_EXAMPLE).unwrap();
-    assert_eq!(RSA_4096_CERT_EXAMPLE.trim_end(), &cert.to_string().unwrap());
+    assert_eq!(
+        RSA_4096_CERT_EXAMPLE.trim_end(),
+        &cert.to_openssh().unwrap()
+    );
 }
 
 #[cfg(all(feature = "ed25519", feature = "fingerprint"))]
